+++ packages/wayfinder-core/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @ar.io/wayfinder-core
 
+## 1.7.1-alpha.2
+
+### Patch Changes
+
+- 3c3c598: Expose `TrustedPeersGatewaysProvider` from index.tsx for top level imports
+
+## 1.7.1-alpha.1
+
+### Patch Changes
+
+- bfea49d: Update npm to use OIDC for publishing
+
+## 1.7.1-alpha.0
+
+### Patch Changes
+
+- 75915fd: Update README.md to include x402 support
+
 ## 1.7.0
 
 ### Minor Changes