+++ packages/wayfinder-react/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @ar.io/wayfinder-react
 
+## 1.0.23-alpha.4
+
+### Patch Changes
+
+- Updated dependencies [3c3c598]
+  - @ar.io/wayfinder-core@1.7.1-alpha.2
+
+## 1.0.23-alpha.3
+
+### Patch Changes
+
+- Updated dependencies [bfea49d]
+  - @ar.io/wayfinder-core@1.7.1-alpha.1
+
+## 1.0.23-alpha.2
+
+### Patch Changes
+
+- Updated dependencies [75915fd]
+  - @ar.io/wayfinder-core@1.7.1-alpha.0
+
 ## 1.0.23
 
 ### Patch Changes